// src/models/patient.ts
export interface Patient {
    key: string;
    responsavel: string;
    equipe: string;
    data: string;
    nome: string;
    idade: string;
    genero: string;
    cor: string;
    profissao: string;
    natural_de: string;
    procedente_de: string;
    queixa: string;
    isda: string;
    pessoal_verminose: string;
    pessoal_dm: string;
    pessoal_has: string;
    pessoal_cancer: string;
    pessoal_tb: string;
    pessoal_avc: string;
    pessoal_alcool: string;
    pessoal_cigarro: string;
    pessoal_drogas: string;
    pessoal_alergias: string;
    pessoal_outros: string;
    pessoal_cirurgias_previas: string;
    familiar_verminose: string;
    familiar_dm: string;
    familiar_has: string;
    familiar_cancer: string;
    familiar_tb: string;
    familiar_avc: string;
    familiar_alergias: string;
    familiar_outros: string;
    exame_fisico_geral_beg: string;
    exame_fisico_geral_reg: string;
    exame_fisico_geral_meg: string;
    exame_fisico_geral_corado: string;
    exame_fisico_geral_descorado: string;
    exame_fisico_geral_descorado_qtd: string;
    exame_fisico_geral_hidratado: string;
    exame_fisico_geral_desidratado: string;
    exame_fisico_geral_desidratado_qtd: string;
    exame_fisico_geral_eupneico: string;
    exame_fisico_geral_dispneico: string;
    exame_fisico_geral_taquipneico: string;
    exame_fisico_geral_acianotico: string;
    exame_fisico_geral_cianose_periferica: string;
    exame_fisico_geral_cianose_central: string;
    exame_fisico_geral_anicterico: string;
    exame_fisico_geral_icterico: string;
    exame_fisico_geral_icterico_grau: string;
    exame_fisico_geral_afebril: string;
    exame_fisico_geral_febril: string;
    exame_fisico_geral_febril_qtd: string;
    exame_fisico_especifico_neuro: string;
    exame_fisico_especifico_ac: string;
    exame_fisico_especifico_fc: string;
    exame_fisico_especifico_pa: string;
    exame_fisico_especifico_ap: string;
    exame_fisico_especifico_fr: string;
    exame_fisico_especifico_abdomen: string;
    exame_fisico_especifico_genitourinario: string;
    exame_fisico_especifico_pele: string;
    conduta: string;
}